Do not check item for null when adding to it
I'm trying to 'fix' broken preconditions being handled
as normal input.

This might be the wrong call here as cJSON is clearly
in general made to handle NULL internally in hopes of the end
user needing to make fewer of his own checks.

But for instance, in this case, a lack of checking
will lead to leaks because this call is supposed to
take ownership and won't if parent is null. So it
seems you need to check for NULL anyways in one
way or another.