Rename cJSON API to sbJSON
The cJSON project has rejected 64-bit integer support for a long
time now, as they want to stick to C89.

In the short time I've spent, mere minutes or hours,
working on this code it's already diverging heavily
so it's best that this becomes a fork with a new name.

'sb' is a meaningless string to the public but it's as
good as any I guess.