Replace flag approach with enum
This is simpler.

If the purpose of the flags was originally to keep the size of the cJSON
struct small, then it was 64 bytes and with or without this
change, it is now 56 bytes after it was changed during the work
to add 64-bit integer support. So we're all good.