Redesign/alter how comparison works
Often for complex objects one wants deep equality,
and this can not be provided by the builtin '=='
operator, which implements only shallow comparison.

That's what sbJSON_Compare is for.

However, cJSON has implemented a hybrid where correct
deep equality is done where possible, but in some cases
where only shallow comparison is possible (and needed),
cJSON intentionally reverses the logic, returning the opposite
results of the builtin shallow comparison operator:
    !(a == b)

Specifically, this occurs when both a and b are NULL.

Without full and proper documentation and justification,
having this in is fully untenable.