Remove case insensitivity - WIP
cJSON started off with being case insensitive,
which is not JSON spec conforming.

They later added case sensitivity as an option
via parameters and additional, longer-named functions.

In this fork we can't have default behavior be
so at odds with the spec.

We could massively rename functions, or just remove support
for case insensitivity. We'll remove support.