We now pass more double tests
When very large numbers had no decimal point, they were parsed as
integers and we couldn't represent them.

But do we even want to? Well it's hard to say. One could imagine
a floating point to string routine which, when trying to keep the
string representation small, would cut off the '.0' part.

Then, as the number gets parsed by us, our ability to handle
the number gets drammatically altered just based on that decision.
That'd be kinda weird. But it's hard to say.